backend/requirements.txt
# Maya Voice SDR Agent - Backend Dependencies
# Python 3.10+ required

# LiveKit Agent Framework
livekit-agents>=1.0.0
livekit-plugins-silero>=1.0.0
livekit-plugins-noise-cancellation>=1.0.0
livekit>=1.0.0

# Environment Management
python-dotenv>=1.0.0

# Async Support
aiohttp>=3.9.0
asyncio-mqtt>=0.16.0

# Logging & Utilities
structlog>=24.0.0
